+++ backend/middlewares/config/config.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	// Port is the port the server will run
+	Port = ":8899"
+)
