--- controller/controller.go
@@ -1,22 +1,0 @@
-package controller
-
-import r "github.com/larien/family-tree/repository"
-
-// Controllers contains all Controllers interface from different domains.
-// When creating a new domain, its interface should be placed here in
-// order to external layers have access to its methods.
-type Controllers struct {
-	Person PersonController
-}
-
-
-// New creates a Controllers layer and shares its reference to the external
-// layer have access to its methods. It should receive as parameter the
-// connections needed by the Controllers layer and the Repository interface.
-// Each Controllers dependency is placed according to each domain.
-func New(repository *r.Repositories) *Controllers {
-	return &Controllers{
-		Person: newPersonController(repository.Person),
-		// implement controllers domains here
-	}
-}
