+++ docker-compose.yml
@@ -1,0 +1,33 @@
+version: "3.2"
+services:
+  neo4j:
+    image: neo4j:latest
+    container_name: "neo4j-db"
+    volumes:
+      - "$HOME/neo4j/data:/data"
+      - "$HOME/neo4j/conf/:/conf/"
+    ports:
+      - 7474:7474
+      - 7687:7687
+    restart: on-failure
+    networks:
+      - neo4j_go_net
+    command: neo4j
+  backend:
+    container_name: "family-tree-go"
+    build: "./backend"
+    ports:
+      - "8080:8080"
+    volumes:
+      - "./backend:/go/src/github.com/larien/family-tree"
+    depends_on:
+      - "neo4j"
+    networks:
+      - neo4j_go_net
+
+networks:
+  neo4j_go_net:
+    driver: bridge
+
+volumes:
+  dgraph:
