--- go.mod
@@ -1,9 +1,0 @@
-module github.com/larien/family-tree
-
-require (
-	github.com/gin-contrib/cors v1.3.0
-	github.com/gin-gonic/gin v1.4.0
-	github.com/neo4j-drivers/gobolt v1.7.4 // indirect
-	github.com/neo4j/neo4j-go-driver v1.7.4
-	github.com/stretchr/testify v1.3.0
-)
