--- middlewares/config/config.go
@@ -1,6 +1,0 @@
-package config
-
-const (
-	// Port is the port the server will run
-	Port = ":8899"
-)
